[refactor] #48 merge conflict 해결
