+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/TimerScreen.kt
@@ -46,6 +46,8 @@ import androidx.compose.ui.unit.dp
 import org.android.bbangzip.R
 import org.android.bbangzip.presentation.common.component.toggle.BbangZipSegmentedButton
 import org.android.bbangzip.presentation.common.util.extension.Gap
+import org.android.bbangzip.presentation.common.util.extension.dropShadow
+import org.android.bbangzip.presentation.common.util.extension.innerShadow
 import org.android.bbangzip.presentation.common.util.extension.noRippleClickable
 import org.android.bbangzip.presentation.ui.shared.SharedContract
 import org.android.bbangzip.presentation.ui.timer.component.CircularProgressBar
@@ -57,10 +59,6 @@ import org.android.bbangzip.presentation.ui.timer.contract.TimerContract
 import org.android.bbangzip.presentation.ui.timer.contract.model.TimerStatus
 import org.android.bbangzip.presentation.ui.timer.contract.model.getTimerFontColor
 import org.android.bbangzip.presentation.ui.timer.contract.model.getTitleText
-import org.android.bbangzip.presentation.util.extension.Gap
-import org.android.bbangzip.presentation.util.extension.dropShadow
-import org.android.bbangzip.presentation.util.extension.innerShadow
-import org.android.bbangzip.presentation.util.extension.noRippleClickable
 import org.android.bbangzip.ui.theme.BBANGZIPANDROIDTheme
 import org.android.bbangzip.ui.theme.BbangZipTheme
