+++ app/src/main/java/org/android/bbangzip/presentation/ui/timer/component/CircularProgressBar.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.android.bbangzip.presentation.util.extension.innerShadow
+import org.android.bbangzip.presentation.common.util.extension.innerShadow
 import org.android.bbangzip.ui.theme.BbangZipTheme
 import org.android.bbangzip.ui.theme.defaultBbangZipBrush
