+++ app/src/main/java/org/android/bbangzip/presentation/ui/todo/bottomsheet/TodoSettingBottomSheet.kt
@@ -24,13 +24,13 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.android.bbangzip.R
-import org.android.bbangzip.presentation.component.bottomsheet.BbangZipBottomSheetSlot
-import org.android.bbangzip.presentation.component.button.BbangZipButtonDefaults
-import org.android.bbangzip.presentation.component.button.BbangzipBaseButton
-import org.android.bbangzip.presentation.component.button.TimeSettingButton
-import org.android.bbangzip.presentation.component.toggle.BbangZipSwitch
+import org.android.bbangzip.presentation.common.component.bottomsheet.BbangZipBottomSheetSlot
+import org.android.bbangzip.presentation.common.component.button.BbangZipButtonDefaults
+import org.android.bbangzip.presentation.common.component.button.BbangzipBaseButton
+import org.android.bbangzip.presentation.common.component.button.TimeSettingButton
+import org.android.bbangzip.presentation.common.component.toggle.BbangZipSwitch
+import org.android.bbangzip.presentation.common.util.extension.Gap
 import org.android.bbangzip.presentation.ui.todo.type.TodoSettingActionType
-import org.android.bbangzip.presentation.util.extension.Gap
 import org.android.bbangzip.ui.theme.BbangZipTheme
 import java.time.LocalTime
